perf: migrate sumtree to rust
As an unfortunate side-effect, this required an
early normalization of datatypes. All idx integers
are now unsigned ints. Fortunately, this should
give a small boost to performance across the board

+++ ReplayTables/_utils/MinMaxHeap.py
@@ -14,9 +14,9 @@ class MinMaxHeap:
             # actual heap values
             np.zeros(4),
             # storage pointers
-            np.zeros(4, dtype=np.uint32),
+            np.zeros(4, dtype=np.int64),
             # back-references
-            np.zeros(4, dtype=np.uint32),
+            np.zeros(4, dtype=np.int64),
         )
 
     def add(self, priority: float, idx: int):
